2 more fixes for the armel build failures
git-svn-id: svn://scribus.net/branches/Version14x/Scribus@17299 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem_textframe.cpp
@@ -554,7 +554,7 @@ struct LineControl {
 		double maxX = colRight - morespace;
 		if (legacy) maxX -= lineCorr;
 
-		double StartX = floor(qMax(line.x, qMin(maxX,breakXPos-maxShrink-1))-1);
+		double StartX = floor(qMax((double)line.x, qMin(maxX,breakXPos-maxShrink-1))-1);
 		int xPos  = static_cast<int>(ceil(maxX));
 
 		QPoint  pt12 (xPos, yAsc);
